+++ src/graphene_elastic/filter_backends/base.py
@@ -17,9 +17,20 @@ __all__ = ("BaseBackend",)
 
 
 class BaseBackend(object):
+    """Base backend."""
 
+    # Prefix for the backend. This is used to isolate the backend variables
+    # but also is used in dynamically constructing the class names of
+    # `graphene` types.
     prefix = None
-    has_fields = False
+
+    # If set to True, an attempt will be made to construct the GraphQL query
+    # fields for backend.
+    has_query_fields = False
+
+    # If set to True, an attempt will be made to construct the connection
+    # fields for backend.
+    has_connection_fields = False
 
     def __init__(self, connection_field, args=None):
         self.connection_field = connection_field
@@ -27,77 +38,50 @@ class BaseBackend(object):
         assert self.prefix
 
     def field_belongs_to(self, field_name):
+        """Check if given filter field belongs to the backend.
+
+        :param field_name:
+        :return:
+        """
         raise NotImplementedError
 
-    def add_arg_prefix(self, arg_name):
-        return "{}_{}".format(self.prefix, arg_name)
-
-    def get_field_name(self, arg_name):
-        if arg_name.startswith(self.prefix):
-            return arg_name.lstrip("{}_".format(self.prefix))
-
-    def arg_belongs_to(self, arg_name):
-        field_name = self.get_field_name(arg_name)
-        if field_name:
-            return self.field_belongs_to(field_name)
-        return False
+    # def add_arg_prefix(self, arg_name):
+    #     return "{}_{}".format(self.prefix, arg_name)
+    #
+    # def get_field_name(self, arg_name):
+    #     """Get field name.
+    #
+    #     :param arg_name:
+    #     :return:
+    #     """
+    #     if arg_name.startswith(self.prefix):
+    #         return arg_name.lstrip("{}_".format(self.prefix))
+    #
+    # def arg_belongs_to(self, arg_name):
+    #     field_name = self.get_field_name(arg_name)
+    #     if field_name:
+    #         return self.field_belongs_to(field_name)
+    #     return False
 
     def filter(self, queryset):
-        raise NotImplementedError
+        """Filter. This method alters current queryset.
 
-    # @property
-    # def filter_fields(self):
-    #     """Filtering filter fields."""
-    #     return getattr(
-    #         self.connection_field.type._meta.node._meta,
-    #         'filter_backend_options',
-    #         {}
-    #     ).get('filter_fields', {})
-    #
-    # @property
-    # def filter_args_mapping(self):
-    #     return {k: k for k, v in self.filter_fields.items()}
-    #
-    # @property
-    # def search_fields(self):
-    #     """Search filter fields."""
-    #     return getattr(
-    #         self.connection_field.type._meta.node._meta,
-    #         'filter_backend_options',
-    #         {}
-    #     ).get('search_fields', {})
-    #
-    # @property
-    # def search_args_mapping(self):
-    #     return {k: k for k, v in self.search_fields.items()}
-    #
-    # @property
-    # def ordering_fields(self):
-    #     return getattr(
-    #         self.connection_field.type._meta.node._meta,
-    #         'filter_backend_options',
-    #         {}
-    #     ).get('ordering_fields', {})
-    #
-    # @property
-    # def ordering_args_mapping(self):
-    #     return {k: k for k, v in self.ordering_fields.items()}
-    #
-    # @property
-    # def ordering_defaults(self):
-    #     return getattr(
-    #         self.connection_field.type._meta.node._meta,
-    #         'filter_backend_options',
-    #         {}
-    #     ).get('ordering_defaults', {})
+        :param queryset:
+        :return:
+        """
+        raise NotImplementedError
 
     @property
     def doc_type(self):
+        """Shortcut to the Elasticsearch document type.
+
+        :return:
+        """
         return self.connection_field.document._doc_type
 
     @classmethod
-    def generic_fields(cls):
-        """Generic backend specific filtering fields.
+    def generic_query_fields(cls):
+        """Generic backend specific query fields.
 
         For instance, for search filter backend it would be
         ``{'search': String()}``.
@@ -108,8 +92,51 @@ class BaseBackend(object):
         return {}
 
     def get_backend_document_fields(self):
-        """Get backend document fields.
+        """Get additional document fields for the backend.
 
+        For instance, the ``Highlight`` backend add additional field named
+        ``highlight`` to the list of fields.
+
+        Sample query:
+
+            query {
+              allPostDocuments(search:{title:{value:"alice"}}) {
+                edges {
+                  node {
+                    id
+                    title
+                    highlight
+                  }
+                }
+              }
+            }
+
+        Sample response:
+
+            {
+              "data": {
+                "allPostDocuments": {
+                  "edges": [
+                    {
+                      "node": {
+                        "id": "UG9zdDp5a1ppVlcwQklwZ2dXbVlJQV91Rw==",
+                        "title": "thus Alice style",
+                        "highlight": {
+                          "title": [
+                            "thus <b>Alice</b> style"
+                          ]
+                        }
+                      }
+                    },
+                    ...
+                  ]
+                }
+              }
+            }
+
+        That ``highlight`` part on both sample query and sample response
+        isn't initially available on the connection level, but added with
+        help of the filter backend.
         :return:
         """
         return OrderedDict()
@@ -121,30 +148,37 @@ class BaseBackend(object):
         """
         raise NotImplementedError
 
-    def get_backend_default_fields_params(self):
-        """Backend default filter params.
+    def get_backend_default_query_fields_params(self):
+        """Get default query fields params for the backend.
 
         :rtype: dict
         :return:
         """
         return {}
 
-    def get_backend_fields(self, items, is_filterable_func, get_type_func):
-        """Construct backend fields.
+    def get_backend_query_fields(self,
+                                 items,
+                                 is_filterable_func,
+                                 get_type_func):
+        """Construct backend query fields.
 
         :param items:
         :param is_filterable_func:
         :param get_type_func:
         :return:
         """
-        params = self.get_backend_default_fields_params()
-        for _k, _v in items:
-            if is_filterable_func(_k):
+        params = self.get_backend_default_query_fields_params()
+        for field, value in items:
+            if is_filterable_func(field):
                 # Getting other backend specific fields (schema dependant)
-                if self.field_belongs_to(_k):
-                    params.update(
-                        {_k: self.get_field_type(_k, _v, get_type_func(_v))}
-                    )
+                if self.field_belongs_to(field):
+                    params.update({
+                        field: self.get_field_type(
+                            field,
+                            value,
+                            get_type_func(value)
+                        )
+                    })
 
         return {
             self.prefix: graphene.Argument(
@@ -159,6 +193,42 @@ class BaseBackend(object):
                 )
             )
         }
+
+    def get_backend_connection_fields_type(self):
+        """Get backend connection fields type.
+
+        Typical use-case - a backend that alters the Connection object
+        and adds additional fields next to `edges` and `pageInfo` (see
+        the `graphene_elastic.relay.connection.Connection` for more
+        information).
+
+        :return:
+        """
+        return {}
+
+    def get_backend_connection_fields(self):
+        """Get backend connection fields.
+
+        Typical use-case - a backend that alters the Connection object
+        and adds additional fields next to `edges` and `pageInfo` (see
+        the `graphene_elastic.relay.connection.Connection` for more
+        information).
+
+        :return:
+        """
+        return {}
+
+    def alter_connection(self, connection, slice):
+        """Alter connection object.
+
+        You can add various properties here, returning the altered object.
+        Typical use-case would be adding facets to the connection.
+
+        :param connection:
+        :param slice:
+        :return:
+        """
+        return connection
 
     @classmethod
     def split_lookup_name(cls, value, maxsplit=-1):